+++ backend/app/commands/api/v1/authenticate_user.rb
@@ -1,0 +1,28 @@
+module Api
+  module V1
+    class AuthenticateUser
+      prepend SimpleCommand
+
+      def initialize(email, password)
+        @email = email
+        @password = password
+      end
+
+      def call
+        JsonWebToken.encode user_id: user.id if user
+      end
+
+      private
+
+      attr_accessor :email, :password
+
+      def user
+        user = User.find_by_email(email)
+        return user if user&.valid_password?(password)
+
+        errors.add :user_authentication, "invalid credentials"
+        nil
+      end
+    end
+  end
+end
