--- backend/app/commands/authenticate_user.rb
@@ -1,24 +1,0 @@
-class AuthenticateUser
-  prepend SimpleCommand
-
-  def initialize(email, password)
-    @email = email
-    @password = password
-  end
-
-  def call
-    JsonWebToken.encode user_id: user.id if user
-  end
-
-  private
-
-  attr_accessor :email, :password
-
-  def user
-    user = User.find_by_email(email)
-    return user if user&.valid_password?(password)
-
-    errors.add :user_authentication, "invalid credentials"
-    nil
-  end
-end
