+++ backend/config/application.rb
@@ -17,6 +17,7 @@ module ConservationInternational
     # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     config.autoload_paths << Rails.root.join("lib")
+    config.eager_load_paths << Rails.root.join("lib")
     config.i18n.available_locales = [:en, :it, :de, :es, :"pt-BR", :fr]
     config.i18n.default_locale = :en
