+++ backend/config/initializers/zeitwerk.rb
@@ -1,0 +1,6 @@
+Rails.autoloaders.main.ignore(
+  "app/admin",
+  "app/assets",
+  "app/javascripts",
+  "app/views"
+)
