+++ backend/spec/support/request_helpers.rb
@@ -1,6 +1,6 @@
 module RequestHelpers
   def auth_token_for(user, password: nil)
-    "Bearer #{AuthenticateUser.call(user.email, password || user.password).result}"
+    "Bearer #{Api::V1::AuthenticateUser.call(user.email, password || user.password).result}"
   end
 
   def response_json
